chore: move out shake utilities into their own module
